Ensure that root flexbox node sizes are floored by their padding border
